mv observer_manager to common_service
