+++ src/common/Cargo.toml
@@ -18,7 +18,6 @@ either = "1"
 futures = { version = "0.3", default-features = false, features = ["alloc"] }
 futures-async-stream = "0.2"
 humantime = "2.1"
-hyper = "0.14"
 itertools = "0.10"
 lazy_static = "1"
 log = "0.4"
@@ -45,8 +44,6 @@ tokio = { version = "=0.2.0-alpha.3", package = "madsim-tokio", features = ["rt"
 tokio-stream = "0.1"
 toml = "0.5"
 tonic = { version = "=0.2.0-alpha.3", package = "madsim-tonic" }
-tower = { version = "0.4", features = ["util", "load-shed"] }
-tower-http = { version = "0.3", features = ["add-extension", "cors"] }
 tracing = { version = "0.1" }
 twox-hash = "1"
 url = "2"
