+++ src/common/src/lib.rs
@@ -54,7 +54,6 @@ pub mod config;
 pub mod field_generator;
 pub mod hash;
 pub mod monitor;
-pub mod service;
 pub mod session_config;
 #[cfg(test)]
 pub mod test_utils;
