+++ src/storage/compactor/Cargo.toml
@@ -9,6 +9,7 @@ clap = { version = "3", features = ["derive"] }
 madsim = "=0.2.0-alpha.3"
 prometheus = { version = "0.13" }
 risingwave_common = { path = "../../common" }
+risingwave_common_service = { path = "../../common/common_service" }
 risingwave_object_store = { path = "../../object_store" }
 risingwave_pb = { path = "../../prost" }
 risingwave_rpc_client = { path = "../../rpc_client" }
